Introduce a new componentreadiness api package to isolate this code
