+++ pkg/api/componentreadiness/component_report_test.go
@@ -1,5 +1,5 @@
 // nolint
-package api
+package componentreadiness
 
 import (
 	"encoding/json"
