+++ pkg/api/utils.go
@@ -52,8 +52,8 @@ func GetPrefixedCacheKey(prefix string, cacheKey interface{}) CacheData {
 	}}
 }
 
-// getDataFromCacheOrGenerate attempts to find a cached record otherwise generates new data.
-func getDataFromCacheOrGenerate[T any](c cache.Cache, cacheOptions cache.RequestOptions, cacheData CacheData, generateFn func() (T, []error), defaultVal T) (T, []error) {
+// GetDataFromCacheOrGenerate attempts to find a cached record otherwise generates new data.
+func GetDataFromCacheOrGenerate[T any](c cache.Cache, cacheOptions cache.RequestOptions, cacheData CacheData, generateFn func() (T, []error), defaultVal T) (T, []error) {
 	if c != nil {
 		cacheKey, err := cacheData.GetCacheKey()
 		if err != nil {
