+++ pkg/sippyserver/server.go
@@ -13,6 +13,7 @@ import (
 	"sync"
 	"time"
 
+	"github.com/openshift/sippy/pkg/api/componentreadiness"
 	"gorm.io/gorm"
 
 	"cloud.google.com/go/storage"
@@ -82,7 +83,7 @@ func NewServer(
 	}
 
 	if bigQueryClient != nil {
-		go api.GetComponentTestVariantsFromBigQuery(bigQueryClient, gcsBucket)
+		go componentreadiness.GetComponentTestVariantsFromBigQuery(bigQueryClient, gcsBucket)
 	}
 
 	return server
@@ -616,7 +617,7 @@ func (s *Server) jsonComponentTestVariantsFromBigQuery(w http.ResponseWriter, re
 		})
 		return
 	}
-	outputs, errs := api.GetComponentTestVariantsFromBigQuery(s.bigQueryClient, s.gcsBucket)
+	outputs, errs := componentreadiness.GetComponentTestVariantsFromBigQuery(s.bigQueryClient, s.gcsBucket)
 	if len(errs) > 0 {
 		log.Warningf("%d errors were encountered while querying test variants from big query:", len(errs))
 		for _, err := range errs {
@@ -639,7 +640,7 @@ func (s *Server) jsonJobVariantsFromBigQuery(w http.ResponseWriter, req *http.Re
 		})
 		return
 	}
-	outputs, errs := api.GetJobVariantsFromBigQuery(s.bigQueryClient, s.gcsBucket)
+	outputs, errs := componentreadiness.GetJobVariantsFromBigQuery(s.bigQueryClient, s.gcsBucket)
 	if len(errs) > 0 {
 		log.Warningf("%d errors were encountered while querying job variants from big query:", len(errs))
 		for _, err := range errs {
@@ -667,7 +668,7 @@ func (s *Server) jsonComponentReportFromBigQuery(w http.ResponseWriter, req *htt
 		})
 		return
 	}
-	allJobVariants, errs := api.GetJobVariantsFromBigQuery(s.bigQueryClient, s.gcsBucket)
+	allJobVariants, errs := componentreadiness.GetJobVariantsFromBigQuery(s.bigQueryClient, s.gcsBucket)
 	if len(errs) > 0 {
 		err := fmt.Errorf("failed to get variants from bigquery")
 		api.RespondWithJSON(http.StatusBadRequest, w, map[string]interface{}{
@@ -677,7 +678,7 @@ func (s *Server) jsonComponentReportFromBigQuery(w http.ResponseWriter, req *htt
 		return
 	}
 	baseRelease, sampleRelease, testIDOption, variantOption, advancedOption, cacheOption, err :=
-		parseComponentReportRequest(s.componentReadinessViews, req, allJobVariants, s.crTimeRoundingFactor)
+		componentreadiness.ParseComponentReportRequest(s.componentReadinessViews, req, allJobVariants, s.crTimeRoundingFactor)
 	if err != nil {
 		api.RespondWithJSON(http.StatusBadRequest, w, map[string]interface{}{
 			"code":    http.StatusBadRequest,
@@ -686,7 +687,7 @@ func (s *Server) jsonComponentReportFromBigQuery(w http.ResponseWriter, req *htt
 		return
 	}
 
-	outputs, errs := api.GetComponentReportFromBigQuery(
+	outputs, errs := componentreadiness.GetComponentReportFromBigQuery(
 		s.bigQueryClient,
 		s.prowURL,
 		s.gcsBucket,
@@ -720,7 +721,7 @@ func (s *Server) jsonComponentReportTestDetailsFromBigQuery(w http.ResponseWrite
 		})
 		return
 	}
-	allJobVariants, errs := api.GetJobVariantsFromBigQuery(s.bigQueryClient, s.gcsBucket)
+	allJobVariants, errs := componentreadiness.GetJobVariantsFromBigQuery(s.bigQueryClient, s.gcsBucket)
 	if len(errs) > 0 {
 		err := fmt.Errorf("failed to get variants from bigquery")
 		api.RespondWithJSON(http.StatusBadRequest, w, map[string]interface{}{
@@ -730,7 +731,7 @@ func (s *Server) jsonComponentReportTestDetailsFromBigQuery(w http.ResponseWrite
 		return
 	}
 	baseRelease, sampleRelease, testIDOption, variantOption, advancedOption, cacheOption, err :=
-		parseComponentReportRequest(s.componentReadinessViews, req, allJobVariants, s.crTimeRoundingFactor)
+		componentreadiness.ParseComponentReportRequest(s.componentReadinessViews, req, allJobVariants, s.crTimeRoundingFactor)
 	if err != nil {
 		api.RespondWithJSON(http.StatusBadRequest, w, map[string]interface{}{
 			"code":    http.StatusBadRequest,
@@ -738,7 +739,7 @@ func (s *Server) jsonComponentReportTestDetailsFromBigQuery(w http.ResponseWrite
 		})
 		return
 	}
-	outputs, errs := api.GetComponentReportTestDetailsFromBigQuery(
+	outputs, errs := componentreadiness.GetComponentReportTestDetailsFromBigQuery(
 		s.bigQueryClient,
 		s.prowURL,
 		s.gcsBucket,
@@ -760,221 +761,6 @@ func (s *Server) jsonComponentReportTestDetailsFromBigQuery(w http.ResponseWrite
 		return
 	}
 	api.RespondWithJSON(http.StatusOK, w, outputs)
-}
-
-func parseVariantOptions(req *http.Request, allJobVariants apitype.JobVariants) (apitype.ComponentReportRequestVariantOptions, error) {
-	var err error
-	variantOption := apitype.ComponentReportRequestVariantOptions{}
-	variantOption.ColumnGroupBy = req.URL.Query().Get("columnGroupBy")
-	variantOption.ColumnGroupByVariants, err = api.VariantsStringToSet(allJobVariants, variantOption.ColumnGroupBy)
-	if err != nil {
-		return variantOption, err
-	}
-	variantOption.DBGroupBy = req.URL.Query().Get("dbGroupBy")
-	variantOption.DBGroupByVariants, err = api.VariantsStringToSet(allJobVariants, variantOption.DBGroupBy)
-	if err != nil {
-		return variantOption, err
-	}
-	variantOption.RequestedVariants = map[string]string{}
-	// Only the dbGroupBy variants can be specifically requested
-	for _, variant := range variantOption.DBGroupByVariants.List() {
-		if value := req.URL.Query().Get(variant); value != "" {
-			variantOption.RequestedVariants[variant] = value
-		}
-	}
-	variantOption.IncludeVariants = req.URL.Query()["includeVariant"]
-	variantOption.IncludeVariantsMap, err = api.IncludeVariantsToMap(allJobVariants, variantOption.IncludeVariants)
-	return variantOption, err
-}
-
-func parseComponentReportRequest(
-	views []apitype.ComponentReportView,
-	req *http.Request,
-	allJobVariants apitype.JobVariants,
-	crTimeRoundingFactor time.Duration) (
-	baseRelease apitype.ComponentReportRequestReleaseOptions,
-	sampleRelease apitype.ComponentReportRequestReleaseOptions,
-	testIDOption apitype.ComponentReportRequestTestIdentificationOptions,
-	variantOption apitype.ComponentReportRequestVariantOptions,
-	advancedOption apitype.ComponentReportRequestAdvancedOptions,
-	cacheOption cache.RequestOptions,
-	err error) {
-
-	// Check if the user specified a view, in which case only some query params can be used.
-	viewRequested := req.URL.Query().Get("view")
-	var view *apitype.ComponentReportView
-	if viewRequested != "" {
-		for i, v := range views {
-			if v.Name == viewRequested {
-				view = &views[i]
-				break
-			}
-		}
-		if view == nil {
-			err = fmt.Errorf("unknown view: %s", viewRequested)
-			return
-		}
-	}
-	if viewRequested != "" {
-		// the following params are not compatible with use of a view and will generate an error if combined with one:
-		if pErr := anyParamSpecified(req,
-			"baseRelease",
-			"sampleRelease",
-			"samplePROrg",
-			"samplePRRepo",
-			"samplePRNumber",
-			"columnGroupBy",
-			"dbGroupBy",
-			"includeVariant",
-			"confidence",
-			"pity",
-			"minFail",
-			"ignoreMissing",
-			"ignoreDisruption",
-		); pErr != nil {
-			err = pErr
-			return
-		}
-		// set params from view
-		variantOption = view.VariantOptions
-		advancedOption = view.AdvancedOptions
-		baseRelease = view.BaseRelease
-		sampleRelease = view.SampleRelease
-	} else {
-		baseRelease.Release = req.URL.Query().Get("baseRelease")
-		if baseRelease.Release == "" {
-			err = fmt.Errorf("missing base_release")
-			return
-		}
-
-		sampleRelease.Release = req.URL.Query().Get("sampleRelease")
-		if sampleRelease.Release == "" {
-			err = fmt.Errorf("missing sample_release")
-			return
-		}
-		// We only support pull request jobs as the sample, not the basis:
-		samplePROrg := req.URL.Query().Get("samplePROrg")
-		samplePRRepo := req.URL.Query().Get("samplePRRepo")
-		samplePRNumber := req.URL.Query().Get("samplePRNumber")
-		if len(samplePROrg) > 0 && len(samplePRRepo) > 0 && len(samplePRNumber) > 0 {
-			sampleRelease.PullRequestOptions = &apitype.PullRequestOptions{
-				Org:      samplePROrg,
-				Repo:     samplePRRepo,
-				PRNumber: samplePRNumber,
-			}
-		}
-
-		variantOption, err = parseVariantOptions(req, allJobVariants)
-		if err != nil {
-			return
-		}
-
-		advancedOption.Confidence = 95
-		confidenceStr := req.URL.Query().Get("confidence")
-		if confidenceStr != "" {
-			advancedOption.Confidence, err = strconv.Atoi(confidenceStr)
-			if err != nil {
-				err = fmt.Errorf("confidence is not a number")
-				return
-			}
-			if advancedOption.Confidence < 0 || advancedOption.Confidence > 100 {
-				err = fmt.Errorf("confidence is not in the correct range")
-				return
-			}
-		}
-
-		advancedOption.PityFactor = 5
-		pityStr := req.URL.Query().Get("pity")
-		if pityStr != "" {
-			advancedOption.PityFactor, err = strconv.Atoi(pityStr)
-			if err != nil {
-				err = fmt.Errorf("pity factor is not a number")
-				return
-			}
-			if advancedOption.PityFactor < 0 || advancedOption.PityFactor > 100 {
-				err = fmt.Errorf("pity factor is not in the correct range")
-				return
-			}
-		}
-
-		advancedOption.MinimumFailure = 3
-		minFailStr := req.URL.Query().Get("minFail")
-		if minFailStr != "" {
-			advancedOption.MinimumFailure, err = strconv.Atoi(minFailStr)
-			if err != nil {
-				err = fmt.Errorf("min_fail is not a number")
-				return
-			}
-			if advancedOption.MinimumFailure < 0 {
-				err = fmt.Errorf("min_fail is not in the correct range")
-				return
-			}
-		}
-
-		advancedOption.IgnoreMissing = false
-		ignoreMissingStr := req.URL.Query().Get("ignoreMissing")
-		if ignoreMissingStr != "" {
-			advancedOption.IgnoreMissing, err = strconv.ParseBool(ignoreMissingStr)
-			if err != nil {
-				err = errors.WithMessage(err, "expected boolean for ignore missing")
-				return
-			}
-		}
-
-		advancedOption.IgnoreDisruption = true
-		ignoreDisruptionsStr := req.URL.Query().Get("ignoreDisruption")
-		if ignoreMissingStr != "" {
-			advancedOption.IgnoreDisruption, err = strconv.ParseBool(ignoreDisruptionsStr)
-			if err != nil {
-				err = errors.WithMessage(err, "expected boolean for ignore disruption")
-				return
-			}
-		}
-	}
-
-	// Params below this point can be used with and without views:
-
-	// TODO: start with a value calculated from the view, then allow overriding:
-	timeStr := req.URL.Query().Get("baseStartTime")
-	baseRelease.Start, err = util.ParseCRReleaseTime(timeStr, crTimeRoundingFactor)
-	if err != nil {
-		err = fmt.Errorf("base start time in wrong format")
-		return
-	}
-	timeStr = req.URL.Query().Get("baseEndTime")
-	baseRelease.End, err = util.ParseCRReleaseTime(timeStr, crTimeRoundingFactor)
-	if err != nil {
-		err = fmt.Errorf("base end time in wrong format")
-		return
-	}
-	timeStr = req.URL.Query().Get("sampleStartTime")
-	sampleRelease.Start, err = util.ParseCRReleaseTime(timeStr, crTimeRoundingFactor)
-	if err != nil {
-		err = fmt.Errorf("sample start time in wrong format")
-		return
-	}
-	timeStr = req.URL.Query().Get("sampleEndTime")
-	sampleRelease.End, err = util.ParseCRReleaseTime(timeStr, crTimeRoundingFactor)
-	if err != nil {
-		err = fmt.Errorf("sample end time in wrong format")
-		return
-	}
-
-	testIDOption.Component = req.URL.Query().Get("component")
-	testIDOption.Capability = req.URL.Query().Get("capability")
-	testIDOption.TestID = req.URL.Query().Get("testId")
-
-	forceRefreshStr := req.URL.Query().Get("forceRefresh")
-	if forceRefreshStr != "" {
-		cacheOption.ForceRefresh, err = strconv.ParseBool(forceRefreshStr)
-		if err != nil {
-			err = errors.WithMessage(err, "expected boolean for force refresh")
-			return
-		}
-	}
-	cacheOption.CRTimeRoundingFactor = crTimeRoundingFactor
-
-	return
 }
 
 func (s *Server) jsonJobBugsFromDB(w http.ResponseWriter, req *http.Request) {
@@ -1904,17 +1690,4 @@ func recordResponse(c cache.Cache, duration time.Duration, w http.ResponseWriter
 
 func (s *Server) GetHTTPServer() *http.Server {
 	return s.httpServer
-}
-
-func anyParamSpecified(req *http.Request, paramName ...string) error {
-	found := []string{}
-	for _, p := range paramName {
-		if req.URL.Query().Get(p) != "" {
-			found = append(found, p)
-		}
-	}
-	if len(found) > 0 {
-		return fmt.Errorf("params cannot be combined with view: %v", found)
-	}
-	return nil
 }
